+++ cborg/tests/Main.hs
@@ -7,11 +7,11 @@ import qualified Tests.UnitTests  as UnitTests
 import qualified Tests.Properties as Properties
 import qualified Tests.Boundary   as Boundary
 import qualified Tests.ByteOffset as ByteOffset
+import qualified Tests.ByteSpan   as ByteSpan
 import qualified Tests.Canonical  as Canonical
 import qualified Tests.Regress    as Regress
 import qualified Tests.UTF8       as UTF8
 import qualified Tests.PreEncoded as PreEncoded
-import qualified Tests.GetInputSpan as GetInputSpan
 
 main :: IO ()
 main = defaultMain tests
@@ -23,7 +23,7 @@ tests =
     , UnitTests.testTree
     , Properties.testTree
     , ByteOffset.testTree
-    , GetInputSpan.testTree
+    , ByteSpan.testTree
     , Boundary.testTree
     , Canonical.testTree
     , Regress.testTree